Add FlowCollectionDelegate
